backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
pymongo==4.6.1
ccxt==4.2.25
python-dotenv==1.0.0
pyyaml==6.0.1
httpx==0.26.0

